fix(location): prevent deletion of a location if it's associated with a FA or a FT #2321
Approved-by: VANDERHAGHEN Luc <luke.vandh@gmail.com>

+++ apps/api/src/location/location.service.ts
@@ -1,7 +1,8 @@
-import { Injectable } from "@nestjs/common";
+import { ForbiddenException, Injectable } from "@nestjs/common";
 import { PrismaService } from "../prisma.service";
 import { CreateLocationRequestDto } from "./dto/create-location.request.dto";
 import { UpdateLocationRequestDto } from "./dto/update-location.request.dto";
+import { IS_NOT_DELETED } from "../common/query/not-deleted.query";
 
 @Injectable()
 export class LocationService {
@@ -26,7 +27,26 @@ export class LocationService {
     });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const [linkedActivities, linkedTasks] = await Promise.all([
+      this.prisma.festivalActivity.findMany({
+        where: { locationId: id, ...IS_NOT_DELETED },
+        select: { id: true },
+      }),
+      this.prisma.festivalTask.findMany({
+        where: { appointmentId: id, ...IS_NOT_DELETED },
+        select: { id: true },
+      }),
+    ]);
+
+    if (linkedActivities.length > 0 || linkedTasks.length > 0) {
+      const activitiesTitles = linkedActivities.map(({ id }) => `FA ${id}`);
+      const tasksTitles = linkedTasks.map(({ id }) => `FT ${id}`);
+      const allLinkedTitles = [...activitiesTitles, ...tasksTitles].join(", ");
+      const errorMessage = `Impossible de supprimer le lieu, il est lié à : ${allLinkedTitles}`;
+      throw new ForbiddenException(errorMessage);
+    }
+
     return this.prisma.signaLocation.delete({ where: { id } });
   }
 }
